rename semaphoreUtils -> SemaphoreUtils
